win32: Add private HR_CHECK() macro
This works like GDK_VK_CHECK() in that it adds a check for a HRESULT and
if the result is not okay, it prints a g_warning() about the statement
that was evaluated and the error message, ideally with its string
version.

This is somewhat brittle because it's unclear if HRESULTs are allowed to
be treated like errors from GetLastError() but that's what everyone else
seems to do, too.