gridview: Rework multirow tile handling
In the presence of sections, we can't really guarantee that big tiles
are rectangular in the sense that they contain MxN items, since sections
might introduce gaps that don't count for n_items.

But we can split off an incomplete first and last row to guarantee that
the first item is in the first column and the last item in the last one.