gridview: Find the correct column for anchors
We use the new helper function to find the column after
estimating how many items from the beginning of the tile
we are.

Note that the offsetting inside the tile is not precise,
since section boundaries might introduce gaps. But once
we've settled on an item, we can determine its columns with
certainty.

The only cases when we still see mismatches is when n_columns
or has_sections changes between get_position_from_allocation
and size_allocate.