gridview: Rewrite get_position_from_allocation
Use the new column helper, and write this function
in a way that will keep working with sections.

The most important point here is that we determine the
column that the anchor item will fall into *after*
other adjustments for position.

This means that the column will be accurate, unless n_columns
or has_sections changes between get_position_for_allocation
and size_allocate.