testsuite: Make sorter generation random
Instead of getting the whole string from every item, get a substring at
offset and length that is randomly chosen when creating the sorter.

This generates way weirder and unpredictable sorting behavior, when each
sorter selects a few random characters to sort by that may or may not be
related between sorters.

The sorters can also have a very high chance of the comparison being
different (when comparing the whole string) or being equal (when
comparing the 4th letter where 3/4 of strings are <= 3 letters).