Add GtkPopoverBin
Adding menus to a widget requires subclassing GtkWidget; adding the
original widget as a child; adding a GtkPopover as an internal child;
and setting a layout manager or handling the popover's layout manually.

GtkPopoverBin is a convenience widget that handles all of these
shenanigans for you. What you're left with is handling when to pop up,
and pop down, the menu.