Add tests for stroking
Add tests to check that any point on a path is
always at most half the line-width away from the
stroke path with that line-width.

Also check that the outlines of random paths look
as expected. We currently have to exclude paths
where points get too close to each other.