textbuffer: Return GObject properties on get_can_redo/undo
It's possible that the textbuffer has been subclassed and
have overridden 'can-redo' or 'can-undo' properties.

gtk_text_buffer_get_can_redo()/undo() should return those
overridden properties in those cases so that the users of the
API (eg: The 'Redo' or 'Undo' button in the text-view context
menu) behaves properly.