Bail out on unzip failures
Downloading something from the GitLab artifact cache will always
"succeed" in getting something, even if that's just an HTML 404 page. We
need to bail out in case we try to unzip that and fail, otherwise we're
going to blow away the references we get from the GTK CI, and just upload
the docs we build inside this branch.